Added more test cases, printing CHS triples now, more explanation comments, formatting

+++ src/read_mbr.h
@@ -1,6 +1,8 @@
 #ifndef READ_MBR_H
 #define READ_MBR_H
 
-int read_mbr(const char* device, char* buf, int buf_size);
+#include <sys/types.h>
+
+off_t read_mbr(const char* device, char* buf, int buf_size);
 
 #endif // READ_MBR_H
